Check for new puzzles when selecting a new one
Each time a puzzle is generated, it is noted down in a file, and subsequent
runs checks the content of that file to avoid generating already-seen puzzles.

To avoid having to find all solutions to a puzzle when comparing, the letters
are sorted before comparing; e.g., if we have previously generated a puzzle
from the word SALTINTAG, we will not allow generating one from ANTASTLIG or
LASTINTAG, since no matter which order the letters are in the posted puzzle,
all three solutions will be valid.